Ping Pong Week 3
display ping pong if the number entered is divisible by 15, pong if it
is divisible by 5, ping if it is divisible by 3, and just display the
number if it does not fulfill any of these requirements.  The program
is able to count into the thousands before the number is finally too
big and crashes (when using the loops/ways understood/known by myself).